+++ src/controllers/component/add.js
@@ -1,0 +1,75 @@
+//External
+require('dotenv').config();
+//Services
+const { addComponentService } = require('../../services/component/add');
+//Enums
+const { statusName, statusDetails } = require('../../enums/database/status');
+const { statusCode } = require('../../enums/http/status-code');
+//Const-vars
+//status-code
+const INTERNAL_SERVER_ERROR_CODE = statusCode.INTERNAL_SERVER_ERROR;
+const BAD_REQUEST_CODE = statusCode.BAD_REQUEST;
+const OK_CODE = statusCode.OK;
+const CONNECTION_ERROR_STATUS = statusName.CONNECTION_ERROR;
+const CONNECTION_ERROR_STATUS_DETAIL = statusDetails.CONNECTION_ERROR_DETAIL;
+const CONNECTION_REFUSED_STATUS = statusName.CONNECTION_REFUSED;
+const CONNECTION_REFUSED_STATUS_DETAIL =
+  statusDetails.CONNECTION_REFUSED_DETAIL;
+// Pagination
+let newComponent;
+let msgResponse;
+let msgLog;
+
+/**
+ * @description add a componente to database
+ * @param {any} req any type
+ * @param {any} res any type
+ * @returns a json object with the transaction performed
+ * @example
+ */
+const addComponentController = async (req, res) => {
+  try {
+    msgResponse = null;
+    msgLog = null;
+    newComponent = await addComponentService(req);
+
+    switch (newComponent) {
+      case CONNECTION_ERROR_STATUS:
+        res
+          .status(INTERNAL_SERVER_ERROR_CODE)
+          .send({ error: CONNECTION_ERROR_STATUS_DETAIL });
+        break;
+      case CONNECTION_REFUSED_STATUS:
+        res
+          .status(INTERNAL_SERVER_ERROR_CODE)
+          .send({ error: CONNECTION_REFUSED_STATUS_DETAIL });
+        break;
+      case 0:
+      case undefined:
+      case null:
+        res
+          .status(BAD_REQUEST_CODE)
+          .send({ error: 'Bad request, could not add a component.' });
+        break;
+      default:
+        if (
+          typeof newComponent === 'object' &&
+          newComponent.hasOwnProperty('id')
+        ) {
+          res.status(OK_CODE).send(newComponent);
+          break;
+        }
+        res.status(BAD_REQUEST_CODE).send({ error: newComponent });
+        break;
+    }
+  } catch (error) {
+    msgResponse = 'ERROR in addComponentController() function.';
+    msgLog = msgResponse + `Caused by ${error}`;
+    console.log(msgLog);
+    res.status(INTERNAL_SERVER_ERROR_CODE).send({ error: msgResponse });
+  }
+};
+
+module.exports={
+    addComponentController
+}
